countdown on the server side

+++ server/src/games/games.gateway.ts
@@ -13,9 +13,12 @@ import { SocketUser } from '../common/types';
 import { HistoriesService } from '../histories/histories.service';
 import { PlayerFinishedDto } from './dto';
 import { GamesService } from './games.service';
+import { calculateTimeLimit } from './utils';
+import { calculateAverageCPs } from '../ranks';
 
 @WebSocketGateway()
 export class GamesGateway implements OnGatewayDisconnect {
+  private readonly gameTimers: Map<string, NodeJS.Timeout> = new Map();
   constructor(
     private readonly gamesService: GamesService,
     private readonly historiesService: HistoriesService,
@@ -33,11 +36,48 @@ export class GamesGateway implements OnGatewayDisconnect {
     const players = await this.gamesService.getPlayersInGame(gameId);
 
     if (players.length === 3) {
-      await this.gamesService.startGame(gameId);
-      this.io.sockets.to(`game:${gameId}`).emit('startGame');
+      const averageCPs = calculateAverageCPs(players);
+      this.startCountdown(gameId, 10, async () => {
+        const game = await this.gamesService.updateTime(gameId, 'startedAt');
+        this.startTimeLimit(
+          gameId,
+          calculateTimeLimit(averageCPs, game.paragraph),
+        );
+        this.io.sockets
+          .to(`game:${gameId}`)
+          .emit('startGame', { startedAt: game.startedAt.toISOString() });
+      });
     }
 
     this.io.sockets.to(`game:${gameId}`).emit('players', players);
+  }
+
+  startCountdown(
+    gameId: number,
+    countdown: number,
+    callback: CallableFunction,
+  ) {
+    const interval = setInterval(() => {
+      this.io.sockets.to(`game:${gameId}`).emit('countdown', countdown);
+      if (countdown === 0) {
+        clearInterval(interval);
+        callback();
+      } else {
+        countdown--;
+      }
+    }, 1000);
+    this.gameTimers.set(`game:${gameId}`, interval);
+  }
+
+  startTimeLimit(gameId: number, countdown: number) {
+    this.startCountdown(gameId, countdown, async () => {
+      // end game here
+      const game = await this.gamesService.updateTime(gameId, 'endedAt');
+      this.io.sockets
+        .to(`game:${gameId}`)
+        .emit('endGame', { endedAt: game.endedAt.toISOString() });
+      this.gameTimers.delete(`game:${gameId}`);
+    });
   }
 
   @UseGuards(WsCookieAuthGuard)
@@ -72,6 +112,13 @@ export class GamesGateway implements OnGatewayDisconnect {
     // check for `gameId` because a player might join in on multiple devices causes NestJS to raise an uncanny exception
     if (!gameId) return;
 
-    await this.gamesService.removeIfEmpty(gameId);
+    const playersInGameCount = await this.gamesService.countPlayersInGame(
+      gameId,
+    );
+    if (playersInGameCount === 0) {
+      await this.gamesService.removeIfEmpty(gameId);
+      clearInterval(this.gameTimers.get(`game:${gameId}`));
+      this.gameTimers.delete(`game:${gameId}`);
+    }
   }
 }
